+++ udata_front/theme/gouvfr/assets/js/api/dataservices.ts
@@ -1,3 +1,6 @@
+import { api, type Dataservice, type Dataset } from "@datagouv/components/ts";
+import type { PaginatedArray } from "./types";
+import { getLocalizedUrl } from "../i18n";
 import { api, type Dataservice, type NewDataservice } from "@datagouv/components/ts";
 import { toValue } from "vue";
 
@@ -16,4 +19,18 @@ export function updateDataservice(dataservice: Dataservice) {
 export async function getDataservice(id: string) {
   const res = await api.get<Dataservice>(`dataservices/${id}/`);
   return res.data;
+}
+
+export async function getOrganizationDataservices(oid: string, page: number, pageSize: number, sortDirection: string) {
+  const resp = await api.get<PaginatedArray<Dataservice>>(getLocalizedUrl(`dataservices/`), {
+    params: { organization: oid, sort: sortDirection, page_size: pageSize, page }
+  });
+  return resp.data;
+}
+
+export async function getUserDataservices(userId: string, page: number, pageSize: number, sortDirection: string) {
+  const resp = await api.get<PaginatedArray<Dataservice>>(getLocalizedUrl(`dataservices/`), {
+    params: { sort: sortDirection, page_size: pageSize, page, owner: userId }
+  });
+  return resp.data;
 }