+++ udata_front/theme/gouvfr/assets/js/components/AdminSidebar/AdminSidebarMenu/AdminSidebarMenu.vue
@@ -46,6 +46,11 @@
           :to="{name: 'me-datasets'}"
           />
           <AdminSidebarLink
+          icon="fr-icon-code-s-slash-line"
+          :label="t('Dataservices')"
+          :to="{name: 'me-dataservices'}"
+          />
+          <AdminSidebarLink
             icon="fr-icon-line-chart-line"
             :label="t('Reuses')"
             :to="{name: 'me-reuses'}"
@@ -61,6 +66,12 @@
             :iconHtml="DatabaseIcon"
             :label="t('Datasets')"
             :to="{name: 'organization-datasets', params: {oid: organization.id}}"
+            @click="$emit('click')"
+          />
+          <AdminSidebarLink
+            icon="fr-icon-code-s-slash-line"
+            :label="t('Dataservices')"
+            :to="{name: 'organization-dataservices', params: {oid: organization.id}}"
             @click="$emit('click')"
           />
           <AdminSidebarLink