+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current (in progress)
 
+- Add beta admin `dataservice` page [#558](https://github.com/datagouv/udata-front/pull/558)
 - Add Dataservice publishing form [#358](https://github.com/datagouv/udata-front/pull/559)
 - Add the me keyword to the rel attribute of the link to the user website [#571](https://github.com/datagouv/udata-front/pull/571)
 - Add dataservice access banner [#573](https://github.com/datagouv/udata-front/pull/573)